Add files via upload
This is the set of files needed to compile the document Garafola_Peter_SCDA_HW1. The files need to be unziped and the individual files placed in a sub-directory called "Data" under the directory containing the Garafola_Peter_SCDA_HW1.RMD file. The document will open and read the files one at a time when it is compiled.